Make build operations available to all smoke tests
We want to start using build operations to track which deprecations are thrown in smoke tests
We first need to make build operations available in the smoke test result
CC already captures build operations, but we update non-CC tests to also capture build operations as well.

We move the CC state loading/store assertions to the result type and out of the smoke test, since build
operation trace is not stored with the result.

We also update the Gradleception builds to use build operations for asserting CC store/load to be more
consistent with the other smoke tests.

Capturing build operation traces is very expensive, which is presumably why the Gradleception builds
did not use it. So, we implement a filtered trace option so we can specify which build operations
to capture. This makes capturing a trace much less costly.