Polish and rename ResolveIvyFactory
The ResolveIvyFactory is an older type that manages the building of ComponentResolver instances for external repositories.
This type no longer is exclusive to ivy, and as such has been renamed to ExternalModuleComponentResolverFactory.

Further, we update this type to no longer require a ResolutionStrategyInternal, and instead pass in the relevant parameters
from the resolution strategy. This decouples the factory from public/legacy types related to Configurations.

We also take this time to clean up artifact resolution queries a bit, which also uses this factory. We no longer create
a detached configuration for each ARQ.