Rework `UserInputHandler` to allow a better interaction with the user.
The API now groups a set of questions into a single batch. This allows the UI to better deal with the user being asked a series of related question, for example in the presence of concurrency.

For now, the implementation simply avoids repeatedly redrawing and erasing the rich console status bars between the questions in batch to get rid of some flicker and avoid the separator between questions being erased. Later, it can apply some concurrency control to avoid interleaving unrelated console output or questions, etc.