Do not completely ignore some files when trying to generate @since information
Previously, this would completely skip parsing Provider. This was OK as long
as we never generated extension methods for this type.

We skipped parsing Provider and a few other classes due to bugs in qdox.

If we did generate an extension for a type that was skipped, FunctionSinceRepository.since(...)
would fail with an unhelpful error and prevent you from running any tests.

Instead of completing ignoring files, we remove the annotation that causes the
qdox bug in a hacky way. In the end, we only needed to skip Transformer and
Provider and not any of the other types.

ApiTypeProvider was changed to have slightly better toString for ApiType, so you can
find things in the map in the debugger.